part 4 done, pending check & cleaning

+++ osa4/bloglist-backend/tests/blogs_api.test.js
@@ -3,17 +3,49 @@ const supertest = require('supertest')
 const app = require('../app')
 //const data = require('./blogs')
 const Blog = require('../models/blog')
-const { newBlog, initialBlogs, newBlog_noLikes, newBlog_noTitle, newBlog_noUrl } = require('./blogs')
+const data = require('./blogs')
+const User = require('../models/user')
+const bcrypt = require('bcrypt')
 //const initialBlogs = data.initialBlogs
 //const newBlog = data.newBlog
 
 const api = supertest(app)
 
+const login = {
+  username: 'tester', password: 'password'
+}
+
+
 beforeEach(async () => {
   await Blog.deleteMany({})
-  const blogObjects = initialBlogs.map(blog => new Blog(blog))
+  await User.deleteMany({})
+
+  const passwordHash = await bcrypt.hash('password', 10)
+  const savedUser = await new User({
+    username: 'tester',
+    name: 'Tester',
+    passwordHash
+  }).save()
+
+  //second user should have no blog posts to their name
+  const passwordHash2 = await bcrypt.hash('wrong', 10)
+  await new User({
+    username: 'badtester',
+    name: 'Bad Tester',
+    passwordHash2
+  }).save()
+
+  data.initialBlogs = data.initialBlogs.map(b => {
+    b.user = savedUser._id
+    return b
+  })
+
+  //console.log(data.initialBlogs)
+
+  const blogObjects = data.initialBlogs.map(blog => new Blog(blog))
   const promiseArray = blogObjects.map(blog => blog.save())
   await Promise.all(promiseArray)
+
 })
 
 
@@ -23,7 +55,7 @@ describe('GET /api/blogs', () => {
       .get('/api/blogs')
       .expect(200)
       .expect('Content-Type', /application\/json/)
-    expect(result.body.length).toBe(initialBlogs.length)
+    expect(result.body.length).toBe(data.initialBlogs.length)
   })
 
   test('All blogs have and "id"', async () => {
@@ -40,35 +72,51 @@ describe('GET /api/blogs', () => {
 
 describe('HTTP POST', () => {
   test('POST to /api/blog creates new blog to db', async () => {
+    const loginResponse = await api.post('/api/login').send(login)
+    const token = loginResponse.body.token
     await api
       .post('/api/blogs')
-      .send(newBlog)
+      .set('Authorization', `Bearer ${token}`)
+      .send(data.newBlog)
       .expect(201)
     const response = await api
       .get('/api/blogs')
       .expect(200)
       .expect('Content-Type', /application\/json/)
-    expect(response.body.length).toBe(initialBlogs.length +1)
+    expect(response.body.length).toBe(data.initialBlogs.length +1)
   })
 
   test('new Blog format should macth intended format', async () => {
+    const loginResponse = await api.post('/api/login').send(login)
+    const token = loginResponse.body.token
     await api
       .post('/api/blogs')
-      .send(newBlog)
+      .set('Authorization', `Bearer ${token}`)
+      .send(data.newBlog)
       .expect(201)
     const response = await api
       .get('/api/blogs')
       .expect(200)
       .expect('Content-Type', /application\/json/)
-    expect(response.body[response.body.length -1]).toMatchObject(newBlog)
+    expect(response.body[response.body.length -1]).toMatchObject(data.newBlog)
+  })
+
+  test('POST fails with 401 if no authorization', async () => {
+    await api
+      .post('/api/blogs')
+      .send(data.newBlog)
+      .expect(401)
   })
 })
 
-describe('default', () => {
+describe('Undefined/missing fields', () => {
   test('"likes" will default to 0 when undefined', async () => {
+    const loginResponse = await api.post('/api/login').send(login)
+    const token = loginResponse.body.token
     await api
       .post('/api/blogs')
-      .send(newBlog_noLikes)
+      .set('Authorization', `Bearer ${token}`)
+      .send(data.newBlog_noLikes)
       .expect(201)
     const response = await api
       .get('/api/blogs')
@@ -78,43 +126,65 @@ describe('default', () => {
   })
 
   test('missing title in blog returns 400 Bad request', async () => {
+    const loginResponse = await api.post('/api/login').send(login)
+    const token = loginResponse.body.token
     await api
       .post('/api/blogs')
-      .send(newBlog_noTitle)
+      .set('Authorization', `Bearer ${token}`)
+      .send(data.newBlog_noTitle)
       .expect(400)
   })
 
   test('missing url in blog returns 400 Bad request', async () => {
+    const loginResponse = await api.post('/api/login').send(login)
+    const token = loginResponse.body.token
     await api
       .post('/api/blogs')
-      .send(newBlog_noUrl)
+      .set('Authorization', `Bearer ${token}`)
+      .send(data.newBlog_noUrl)
       .expect(400)
   })
 })
 
 describe('DELETE', () => {
   test('API should return correct amount of blogs after DELETE', async () => {
+    const loginResponse = await api.post('/api/login').send(login)
+    const token = loginResponse.body.token
+
     const resp = await api.get('/api/blogs')
     const id = resp.body[0].id
-    await api.delete(`/api/blogs/${id}`).expect(204)
+    await api.delete(`/api/blogs/${id}`).set('Authorization', `Bearer ${token}`).expect(204)
 
     const resp2 = await api
       .get('/api/blogs')
       .expect(200)
       .expect('Content-Type', /application\/json/)
-    expect(resp2.body.length).toBe(initialBlogs.length -1)
+    expect(resp2.body.length).toBe(data.initialBlogs.length -1)
   })
 
   test('Incorrect or missing id should return error', async () => {
+    const loginResponse = await api.post('/api/login').send(login)
+    const token = loginResponse.body.token
+
     //400 Bad request
-    await api.delete('/api/blogs/test').expect(400)
+    await api.delete('/api/blogs/test').set('Authorization', `Bearer ${token}`).expect(400)
     //404 Not found
-    await api.delete('/api/blogs/').expect(404)
+    await api.delete('/api/blogs/').set('Authorization', `Bearer ${token}`).expect(404)
     const response = await api
       .get('/api/blogs')
       .expect(200)
       .expect('Content-Type', /application\/json/)
-    expect(response.body.length).toBe(initialBlogs.length)
+    expect(response.body.length).toBe(data.initialBlogs.length)
+  })
+
+  test('delete without or with wrong authorization returns 401', async () => {
+    const resp = await api.get('/api/blogs')
+    const id = resp.body[0].id
+    await api.delete(`/api/blogs/${id}`).expect(401)
+
+    const loginResponse = await api.post('/api/login').send({username: 'badtester', password: 'wrong'})
+    const token = loginResponse.body.token
+    await api.delete(`/api/blogs/${id}`).set('Authorization', `Bearer ${token}`).expect(401)
   })
 
 })
@@ -130,29 +200,29 @@ describe('UPDATE', () => {
       .get('/api/blogs')
       .expect(200)
       .expect('Content-Type', /application\/json/)
-    expect(resp2.body.length).toBe(initialBlogs.length)
+    expect(resp2.body.length).toBe(data.initialBlogs.length)
   })
 
   test('Incorrect or missing id should return error', async () => {
-    await api.put('/api/blogs/test').send(newBlog).expect(400)
-    await api.put('/api/blogs/').send(newBlog).expect(404)
+    await api.put('/api/blogs/test').send(data.newBlog).expect(400)
+    await api.put('/api/blogs/').send(data.newBlog).expect(404)
     const response = await api
       .get('/api/blogs')
       .expect(200)
       .expect('Content-Type', /application\/json/)
-    expect(response.body.length).toBe(initialBlogs.length)
+    expect(response.body.length).toBe(data.initialBlogs.length)
   })
 
   test('API should return correct blog after UPDATE', async () => {
     const resp = await api.get('/api/blogs')
     const blog = resp.body[0]
-    await api.put(`/api/blogs/${blog.id}`).send(newBlog).expect(200)
+    await api.put(`/api/blogs/${blog.id}`).send(data.newBlog).expect(200)
 
     const resp2 = await api
       .get(`/api/blogs/${blog.id}`)
       .expect(200)
       .expect('Content-Type', /application\/json/)
-    expect(resp2.body).toMatchObject(newBlog)
+    expect(resp2.body).toMatchObject(data.newBlog)
   })
 
 })
